Renamed and moved backtest functions into backtest.py Initial implementation of 'run' method
