+++ optopsy/__init__.py
@@ -1,4 +1,4 @@
 from .data import get as get
 from .data import gets as gets
-from .backtest import simulate
 from .enums import *
+from .backtest import run
