+++ optopsy/backtest.py
@@ -1,3 +1,67 @@
+from .enums import Period
+from .option_queries import opt_type
+from functools import reduce
+import optopsy.filters as fil
+import pandas as pd
+
+sort_by = ['underlying_symbol', 'quote_date', 'option_type', 'expiration', 'strike']
+on = ['underlying_symbol', 'option_type', 'expiration', 'strike']
+
+default_entry_filters = {
+    "std_expr": False,
+    "contract_size": 10,
+    "entry_dte": Period.FOUR_WEEKS.value,
+    "exit_dte": None
+}
+
+
+def _create_legs(data, leg):
+    return (
+        data
+            .pipe(opt_type, option_type=leg[0])
+            .assign(ratio=leg[1])
+    )
+
+
+def _apply_filters(legs, filters):
+    if not filters:
+        return legs
+    else:
+        return [reduce(lambda l, f: getattr(fil, f)(l, filters[f], idx), filters, leg)
+            for idx, leg in enumerate(legs)]
+
+
+def _filter_data(data, filters):
+    data = data if isinstance(data, list) else [data]
+    return pd.concat(_apply_filters(data, entry_filters)).sort_values(sort_by)
+    
+
+def create_spread(data, leg_structs, filters):
+    legs = [_create_legs(data, leg) for leg in leg_structs]
+
+    # merge and apply leg filters to create spread
+    filters = {**default_entry_filters, **filters}
+    entry_filters = {f: filters[f]
+        for f in filters if (not f.startswith('entry_spread') and 
+										not f.startswith('exit_'))}
+    spread = _filter_data(legs, entry_filters)
+
+    # apply spread level filters to spread
+    spread_filters = {f: filters[f] for f in filters if f.startswith('entry_spread')}
+    return _filter_data(spread, spread_filters)
+
+
 # this is the main function that runs the backtest engine
-def simulate(data, trades, filters):
-    pass
+def run(data, trades, filters):
+
+    trades = trades if isinstance(trades, list) else [trades]
+    
+    #merge trades from multiple underlyings if applicable
+    all_trades = pd.concat(trades).sort_values(sort_by)
+
+    # for each option to be traded, determine the historical price action
+    exit_filters = {f: filters[f] for f in filters if f.startswith('exit_')}
+    results = pd.merge(all_trades, data, on=on).pipe(_filter_data, exit_filters)
+    
+    
+
