+++ optopsy/option_strategies.py
@@ -1,5 +1,5 @@
 from .enums import OrderAction, OptionType
-from .option_spreads import create_spread
+from .backtest import create_spread
 
 
 def _add_date_range(s, e, f):
